+++ crates/core/utils/src/lib.rs
@@ -9,6 +9,7 @@ pub mod shell;
 pub mod time;
 pub mod yaml;
 
+pub use async_trait::async_trait;
 pub use lazy_static::lazy_static;
 
 use cached::proc_macro::cached;